+++ R/online.r
@@ -1,17 +1,17 @@
-#' Search and retrieve toxicity records from the on-line database
+#' Search and retrieve toxicity records from the online database
 #'
-#' `r lifecycle::badge('experimental')` Functions to search and retrieve records from the on-line database at
+#' `r lifecycle::badge('experimental')` Functions to search and retrieve records from the online database at
 #' <https://cfpub.epa.gov/ecotox/search.cfm>.
 #'
-#' The functions described here to search and retrieve records from the on-line database are experimental. This is because this feature is
-#' not formally supported by the EPA, and it may break in future iterations of the on-line database. The functions form an interface between
+#' The functions described here to search and retrieve records from the online database are experimental. This is because this feature is
+#' not formally supported by the EPA, and it may break in future iterations of the online database. The functions form an interface between
 #' R and the ECOTOX website and is therefore limited by its restrictions as described in the package documentation: [ECOTOXr::ECOTOXr]. The
 #' functions should therefore be used with caution.
 #' 
 #' @note **IMPORTANT:** when you plan to perform multiple adjacent searches (for instance in a loop), please insert a call to [Sys.sleep()].
 #' This to avoid overloading the server and getting your IP address banned from the server.
 #'
-#' @param fields A named `list` of `character`s, used to build a search for for the on-line search query of
+#' @param fields A named `list` of `character`s, used to build a search for for the online search query of
 #' <https://cfpub.epa.gov/ecotox/search.cfm>.
 #' Use [list_ecotox_web_fields()] to construct a valid list.
 #' @param habitat Use `aquire` (default) to retrieve aquatic data, `terrestrial` for, you've guessed it, terrestrial data.
@@ -83,9 +83,9 @@ websearch_ecotox <- function(
   if (!grepl("spreadsheet", httr_result$headers$`content-type`)) {
     warn_text <- httr_result |> httr2::resp_body_html() |> rvest::html_text2() |>
       stringr::str_replace("Warning", "") |> trimws()
-    warn_text <- paste(warn_text, "Returning on-line preview data only")
+    warn_text <- paste(warn_text, "Returning online preview data only")
     warning(warn_text)
-    return(list(`On-line preview` = table_preview))
+    return(list(`online preview` = table_preview))
   }
   
   ## otherwise return full data
@@ -101,7 +101,7 @@ websearch_ecotox <- function(
         names = sheet_names
       )
     )
-  data_tables[["On-line preview"]] <- table_preview
+  data_tables[["online preview"]] <- table_preview
   unlink(tab_file)
   
   return(data_tables)
@@ -130,12 +130,12 @@ list_ecotox_web_fields <- function(...) {
 #' `r lifecycle::badge('experimental')` Search <https://comptox.epa.gov/dashboard> for substances and their chemico-physical properties
 #' and meta-information.
 #'
-#' The [CompTox Chemicals Dashboard](https://comptox.epa.gov/dashboard) is a freely accessible on-line U.S. EPA database.
+#' The [CompTox Chemicals Dashboard](https://comptox.epa.gov/dashboard) is a freely accessible online U.S. EPA database.
 #' It contains information on physico-chemical properties, environmental fate and transport, exposure, usage, *in vivo* toxicity,
 #' and *in vitro* bioassay of a wide range of substances.
 #' 
-#' The function described here to search and retrieve records from the on-line database is experimental. This is because this feature is
-#' not formally supported by the EPA, and it may break in future incarnations of the on-line database. The function forms an interface between
+#' The function described here to search and retrieve records from the online database is experimental. This is because this feature is
+#' not formally supported by the EPA, and it may break in future incarnations of the online database. The function forms an interface between
 #' R and the [CompTox](https://comptox.epa.gov/dashboard) website and is therefore limited by the restrictions documented there.
 #' @param searchItems A `vector` of `character`s where each element is a substance descriptor (any of the selected `identifierType`s) you
 #' wish to query.