+++ R/process.r
@@ -76,7 +76,9 @@ process_ecotox_numerics <- function(x, .fns = as_numeric_ecotox, ..., .names = N
 #'  * If there is a hyphen present (not preceded by an "`"e"` or `"E"`) it is probably
 #'    representing a range of values. When `range_fun` is `NULL` it will result in a `NA`.
 #'    Otherwise, the numbers are split at the hyphen and aggregated with `range_fun`
-#'  
+#' 
+#' It is your own responsibility to check if the sanitising steps are appropriate for
+#' your analyses.
 #' @param x A vector of `character` strings. It expects fields as commonly returned
 #' from the ECOTOX database.
 #' @param range_fun Function to summarise range values. If `NULL` range values are